Move GooglePayMark to own file
