+++ packages/wpcom-checkout/src/payment-methods/google-pay.tsx
@@ -13,6 +13,7 @@ import type { Stripe, StripeConfiguration } from '@automattic/calypso-stripe';
 import PaymentRequestButton from '../payment-request-button';
 import { usePaymentRequestOptions, useStripePaymentRequest } from './web-pay-utils';
 import { PaymentMethodLogos } from '../payment-method-logos';
+import { GooglePayMark } from '../google-pay-mark';
 
 const debug = debugFactory( 'wpcom-checkout:google-pay-payment-method' );
 
@@ -95,39 +96,4 @@ export function GooglePaySubmitButton( {
 
 export function GooglePaySummary(): JSX.Element {
 	return <React.Fragment>{ 'Google Pay' }</React.Fragment>;
-}
-
-export function GooglePayMark( { fill }: { fill: string } ): JSX.Element {
-	return (
-		<svg viewBox="0 0 524 206" fill="none" xmlns="http://www.w3.org/2000/svg">
-			<path
-				d="M244.6 100.2V160.7H225.4V11.3H276.3C289.2 11.3 300.2 15.6 309.2 24.2C318.4 32.8 323 43.3 323 55.7C323 68.4 318.4 78.9 309.2 87.4C300.3 95.9 289.3 100.1 276.3 100.1H244.6V100.2ZM244.6 29.7V81.8H276.7C284.3 81.8 290.7 79.2 295.7 74.1C300.8 69 303.4 62.8 303.4 55.8C303.4 48.9 300.8 42.8 295.7 37.7C290.7 32.4 284.4 29.8 276.7 29.8H244.6V29.7Z"
-				fill={ fill }
-			/>
-			<path
-				d="M373.2 55.1C387.4 55.1 398.6 58.9 406.8 66.5C415 74.1 419.1 84.5 419.1 97.7V160.7H400.8V146.5H400C392.1 158.2 381.5 164 368.3 164C357 164 347.6 160.7 340 154C332.4 147.3 328.6 139 328.6 129C328.6 118.4 332.6 110 340.6 103.8C348.6 97.5 359.3 94.4 372.6 94.4C384 94.4 393.4 96.5 400.7 100.7V96.3C400.7 89.6 398.1 84 392.8 79.3C387.5 74.6 381.3 72.3 374.2 72.3C363.5 72.3 355 76.8 348.8 85.9L331.9 75.3C341.2 61.8 355 55.1 373.2 55.1ZM348.4 129.3C348.4 134.3 350.5 138.5 354.8 141.8C359 145.1 364 146.8 369.7 146.8C377.8 146.8 385 143.8 391.3 137.8C397.6 131.8 400.8 124.8 400.8 116.7C394.8 112 386.5 109.6 375.8 109.6C368 109.6 361.5 111.5 356.3 115.2C351 119.1 348.4 123.8 348.4 129.3Z"
-				fill={ fill }
-			/>
-			<path
-				d="M523.5 58.4L459.5 205.6H439.7L463.5 154.1L421.3 58.4H442.2L472.6 131.8H473L502.6 58.4H523.5Z"
-				fill={ fill }
-			/>
-			<path
-				d="M168.23 88C168.23 81.74 167.67 75.75 166.63 69.99H86.15V102.99L132.5 103C130.62 113.98 124.57 123.34 115.3 129.58V150.99H142.89C159 136.08 168.23 114.04 168.23 88Z"
-				fill="#4285F4"
-			/>
-			<path
-				d="M115.31 129.58C107.63 134.76 97.74 137.79 86.17 137.79C63.82 137.79 44.86 122.73 38.07 102.43H9.61V124.51C23.71 152.49 52.69 171.69 86.17 171.69C109.31 171.69 128.75 164.08 142.9 150.98L115.31 129.58Z"
-				fill="#34A853"
-			/>
-			<path
-				d="M35.39 86.05C35.39 80.35 36.34 74.84 38.07 69.66V47.58H9.61C3.78 59.15 0.500015 72.21 0.500015 86.05C0.500015 99.89 3.79 112.95 9.61 124.52L38.07 102.44C36.34 97.26 35.39 91.75 35.39 86.05Z"
-				fill="#FABB05"
-			/>
-			<path
-				d="M86.17 34.3C98.8 34.3 110.11 38.65 119.04 47.15L143.49 22.72C128.64 8.89 109.28 0.399994 86.17 0.399994C52.7 0.399994 23.71 19.6 9.61 47.58L38.07 69.66C44.86 49.36 63.82 34.3 86.17 34.3Z"
-				fill="#E94235"
-			/>
-		</svg>
-	);
 }
